feat: Get variable values as `int`, `double` and `bool` (#65)
Allowes to get the variable values casted to the specific type. If the variable does not exist the `fallback` is used. However when also no `fallback` is provided an `AssertionError` is thrown. Furthermore a `FormatException` is thrown when the variable exists but the value can not be casted to the specific type.